requirements.txt
certifi==2019.3.9
chardet==3.0.4
idna==2.8
influxdb==5.2.2
python-dateutil==2.8.0
pytz==2019.1
PyYAML==5.1
requests==2.21.0
six==1.12.0
urllib3==1.24.1
